fix(interceptor): cast total_lost to i32 for RFC 3550 compliance
Follows the total_lost type change in rtc-rtcp (fix/rtcp-sctp-correctness).
The internal accumulator field stays u32; cast at assignment to match the
signed 24-bit integer field in ReceptionReport per RFC 3550 §6.4.1.

+++ rtc-interceptor/src/report/receiver_stream.rs
@@ -169,7 +169,7 @@ impl ReceiverStream {
                     | (self.last_seq_num as u32 & 0xFFFF),
                 last_sender_report: self.last_sender_report,
                 fraction_lost,
-                total_lost: self.total_lost,
+                total_lost: self.total_lost as i32,
                 delay,
                 jitter: self.jitter as u32,
             }],
